* commit '7454ffecd5141adedb66adea0489f2441f3cca9b': (30 commits)
  BDS-1188: Drop DimensionFilterTest
  BDS-1188: Get rid of nullValue in RawDimensions.
  BDS-1188: Fix according to comments
  BDS-1188: Make DimOrdering serializable
  BDS-1188: Set proper type in InMemoryExternalLinkBase
  BDS-1188: Fix failed test for request handler
  BDS-1188: Handle dimension datatype in upserts, better error messages.
  BDS-1188: Drop unused code
  BDS-1188: Fix warning
  BDS-1188: Fix put and test for it
  BDS-1188: add missing header
  BDS-1188: Add shorts and bytes support to data frame converter
  BDS-1188: Fix compilation with Scala 2.11
  BDS-1188: Make all tests but one work. Optimize constants, see BDS-785 part 1.
  BDS-1188: Make all TsdDaoHbase tests pass
  BDS-1188: Support byte an short, fix a number of tests
  BDS-1188: Make external link tests work
  BDS-1188: Make core tests pass
  BDS-1188: Another bunch of tests to compile
  BDS-1188: Make core tests compile
  ...

+++ yupana-jdbc/src/test/scala/org/yupana/jdbc/YupanaResultSetTest.scala
@@ -265,7 +265,7 @@ class YupanaResultSetTest extends FlatSpec with Matchers with MockFactory {
         DataType[Boolean],
         DataType[Int],
         DataType[String],
-        DataType[Double],
+        DataType[BigDecimal],
         DataType[Long],
         DataType[BigDecimal]
       ),
@@ -493,7 +493,7 @@ class YupanaResultSetTest extends FlatSpec with Matchers with MockFactory {
     val result = SimpleResult(
       "test",
       Seq("int", "string", "double", "time"),
-      Seq(DataType[Int], DataType[String], DataType[Double], DataType[Time]),
+      Seq(DataType[Int], DataType[String], DataType[BigDecimal], DataType[Time]),
       Iterator(
         Array[Option[Any]](Some(42), Some("foo"), None, Some(Time(1234567L)))
       )
